+++ base/deprecated.jl
@@ -1584,7 +1584,7 @@ for op in (:exp, :exp2, :exp10, :log, :log2, :log10,
     @eval @deprecate ($op)(x::AbstractSparseVector{<:Number,<:Integer}) ($op).(x)
 end
 
-# deprecate reamining vectorized methods from Base.Dates
+# deprecate remaining vectorized methods from Base.Dates
 @eval Dates @deprecate(
     DateTime(Y::AbstractArray{<:AbstractString}, f::AbstractString; locale::Locale=ENGLISH),
     DateTime.(Y, f; locale=locale) )
@@ -1597,6 +1597,12 @@ end
 @eval Dates @deprecate(
     Date(Y::AbstractArray{<:AbstractString}, df::DateFormat=ISODateFormat),
     Date.(Y, df) )
+@eval Dates @deprecate(
+    format(Y::AbstractArray{<:TimeType}, f::AbstractString; locale::Locale=ENGLISH),
+    format.(Y, f; locale=locale) )
+@eval Dates @deprecate(
+    format(Y::AbstractArray{T}, df::DateFormat=default_format(T)) where {T<:TimeType},
+    format.(Y, df) )
 
 # PR #22182
 @deprecate is_apple   Sys.isapple