+++ test/dates/io.jl
@@ -328,8 +328,7 @@ dr2 = [Dates.Date(2000) : Dates.Date(2000, 1, 10);]
 @test Dates.DateTime.(dr) == Dates.DateTime.(dr2)
 @test Dates.DateTime.(dr, "yyyy-mm-dd") == Dates.DateTime.(dr2)
 
-@test Dates.format(dr2) == dr
-@test Dates.format(dr2, "yyyy-mm-dd") == dr
+@test Dates.format.(dr2, "yyyy-mm-dd") == dr
 
 @test typeof(Dates.Date.(dr)) == Array{Date, 1}