+++ NEWS.md
@@ -299,8 +299,8 @@ Deprecated or removed
   * Calling `union` with no arguments is deprecated; construct an empty set with an appropriate
     element type using `Set{T}()` instead ([#23144]).
 
-  * Vectorized `DateTime` and `Date` methods have been deprecated in favor of dot-syntax
-    ([#CATS]).
+  * Vectorized `DateTime`, `Date`, and `format` methods have been deprecated in favor of
+    dot-syntax ([#23207]).
 
   * `Base.cpad` has been removed; use an appropriate combination of `rpad` and `lpad`
     instead ([#23187]).
@@ -1160,3 +1160,4 @@ Command-line option changes
 [#23144]: https://github.com/JuliaLang/julia/issues/23144
 [#23157]: https://github.com/JuliaLang/julia/issues/23157
 [#23187]: https://github.com/JuliaLang/julia/issues/23187
+[#23207]: https://github.com/JuliaLang/julia/issues/23207